% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycvb..R
\name{cycvb}
\alias{cycvb}
\title{cycvb}
\usage{
cycvb(x, noise.type = c("gaussian", "t", "gumbel", "laplace"), seed = 1)
}
\arguments{
\item{x}{The observed data following a structural equation model}

\item{noise.type}{Type of the distribution of the noise in the structural equation model. It needs to be one of the following: 'gaussian','t','laplace','gumbel'.}

\item{seed}{Set a seed to reproduce results. Default:1}
}
\value{
Inclusion probability of the adjacent matrix
}
\description{
cycvb
}
\examples{
cycvb(matrix(rt(500*10,3),500,10),"t")
}
