% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad}
\alias{grad}
\title{grad}
\usage{
grad(x, A, mu_u, mu_s, mu_v, noiseType, noiseParams, trans_u, n_randomSample)
}
\arguments{
\item{x}{The observed data following a structural equation model}

\item{A}{Parameter a of the hard concrete distribution}

\item{mu_u}{U matrix of the SVD of matrix mu}

\item{mu_s}{Diagnal matrix of the SVD of matrix mu}

\item{mu_v}{V matrix of the SVD of matrix mu}

\item{noiseType}{Type of the distribution of the noise in the structural equation model. It needs to be one of the following: 'gaussian','t','laplace','gumbel'.}

\item{noiseParams}{Parameter of noise distribution}

\item{trans_u}{Monte Carlo sample of a transformation of uniform distribution}

\item{n_randomSample}{Monte Carlo sample of normal distribution}
}
\value{
gradient of ELBO funcion
}
\description{
grad
}
